cilium, datapath: don't upsert aws-cni vlan routes
When running in chaining mode, AWS creates a vlan for endpoints with an SG attached and intentionally
avoids upserting a route for this endpoint so that traffic flows over the VPC, outside of the node.
Cilium, in it's normal functionality, upserts that route into the main routing table causing
asymetric routing and the packets to be dropped. By avoiding upserting that route, the traffic can flow
the way it was intended to flow.